% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_board.R
\name{add_board}
\alias{add_board}
\title{Add board in the correct order}
\usage{
add_board(
  project_id,
  name = "Development",
  labels_order = c("Bloqué", "Prêt", "En cours", "Révision", "Pré-validation",
    "A valider"),
  lg = "fr"
)
}
\arguments{
\item{project_id}{project_id}

\item{name}{Name of the Board}

\item{labels_order}{Name of the labels already existing in the project}

\item{lg}{language of the board, "fr" or "en"}
}
\description{
Add board in the correct order
}
\examples{
\dontrun{
add_board(
  project_id = project_id
)
#'}
}
