% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_board_github.R
\name{add_board_github}
\alias{add_board_github}
\title{add_board_github}
\usage{
add_board_github(
  owner,
  repo,
  columns = c("Open", "Blocked", "Meta", "Ready", "In Progress", "Review", "Validation")
)
}
\arguments{
\item{owner}{Owner of the repository}

\item{repo}{Repository name}

\item{columns}{Names of columns to create in the Board}
}
\description{
add_board_github
}
\examples{
\dontrun{
add_board_github(
  owner = "ghowner",
  repo = "areponame"
)
}
}
