% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_git_templates.R
\name{add_git_templates}
\alias{add_git_templates}
\title{Add MR GitLab or GitHub template and local issue template}
\usage{
add_git_templates(
  project_path = ".",
  type = c("commit", "mr"),
  target_dir = ".gitlab"
)
}
\arguments{
\item{project_path}{project_path}

\item{type}{type}

\item{target_dir}{Directory name where to save templates}
}
\value{
Side effect: template in .gitlab/.github and in .git
}
\description{
Add MR GitLab or GitHub template and local issue template
}
\examples{
\dontrun{
add_git_templates(
  project_path = project_path,
  type = c("commit", "mr")
)
}
}
