% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{add_issue_clients}
\alias{add_issue_clients}
\title{Add First issue client for GitLab}
\usage{
add_issue_clients(project_id, project_name, group_url)
}
\arguments{
\item{project_id}{project_id}

\item{project_name}{project_name}

\item{group_url}{group_url}
}
\value{
A tibble with the issue added and GitLab infos
}
\description{
Add First issue client for GitLab
}
\examples{
\dontrun{
add_issue_clients(
  project_id = "<get_your_id_project>",
  project_name = "<get_your_project_name>",
  group_url = "<group_url_repo>" # should looks like "https://gitlab.com/cervan.girard/"
)
}
}
