% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{add_issue_dev}
\alias{add_issue_dev}
\title{Create issue for developers to initiate the project}
\usage{
add_issue_dev(project_id, project_name, group_url)
}
\arguments{
\item{project_id}{project_id}

\item{project_name}{project_name}

\item{group_url}{group_url}
}
\value{
Side effect: new issue on GitLab
}
\description{
Create issue for developers to initiate the project
}
\examples{
# exemple nécessite des opérations sur gitlab : à voir
\dontrun{
add_issue_dev(
  project_id = "<get_your_id_project>",
  project_name = "<get_your_project_name>",
  group_url = "<group_url_repo>" # should looks like "https://gitlab.com/cervan.girard/"
)
}
}
