% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{add_issue_dev_github}
\alias{add_issue_dev_github}
\title{Create issue for dev on github}
\usage{
add_issue_dev_github(owner, repo)
}
\arguments{
\item{owner}{owner of the project}

\item{repo}{name of the project}
}
\value{
tibble with github issue info
}
\description{
Create dev first issue in github
}
\examples{
\dontrun{
add_issue_dev_github(
  owner = "owner",
  repo = "repo"
)
}
}
