% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{add_issue_kickoff}
\alias{add_issue_kickoff}
\title{Add Kickoff issue client for GitLab}
\usage{
add_issue_kickoff(project_id)
}
\arguments{
\item{project_id}{project_id}
}
\value{
A tibble with the issue added and GitLab infos
}
\description{
Add Kickoff issue client for GitLab
}
\examples{
\dontrun{
add_issue_kickoff(project_id = "<get_your_id_project>")
}
}
