% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{add_issue_kickoff_github}
\alias{add_issue_kickoff_github}
\title{Add Kickoff issue for GitHub}
\usage{
add_issue_kickoff_github(owner, repo)
}
\arguments{
\item{owner}{GitHub owner (username or organisation)}

\item{repo}{Name of the repository}
}
\value{
A tibble with the issue added and GitHub infos
}
\description{
Add Kickoff issue for GitHub
}
\examples{
\dontrun{
add_issue_kickoff_github(
  owner = "owner",
  repo = "repo"
)
}
}
