% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kit_package.R
\name{add_kit_package}
\alias{add_kit_package}
\title{Add files necessary for a package, including {golem}}
\usage{
add_kit_package(
  project_path = ".",
  type = c("package", "renv"),
  pkgdown_yml = NULL
)
}
\arguments{
\item{project_path}{project_path}

\item{type}{Type of dev_history. Multiple types possible among "package", "book", "renv".}

\item{pkgdown_yml}{path to yaml conf for pkgdown}
}
\value{
Side Effect, add new files :
\itemize{
\item dev/dev_history
\item templates for pkgdown if pkgodwn_yml is not null
\item testthat
}
}
\description{
Add files necessary for a package, including {golem}
}
\examples{
withr::with_tempdir({
  project_path <- getwd()
  usethis::create_package(path = project_path, open = FALSE)
  add_kit_package(project_path, type = c("package", "book", "renv"))
})
}
