% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kit_project.R
\name{add_kit_project}
\alias{add_kit_project}
\title{Add necessary files for any R project}
\usage{
add_kit_project(project_path = ".", name_licence, type_licence)
}
\arguments{
\item{project_path}{project_path}

\item{name_licence}{Name for the licence}

\item{type_licence}{should be a function, example : type_licence = usethis::use_proprietary_license}
}
\value{
Side Effect, add new files :
\itemize{
\item gitattributes
\item NEWS.md
}
}
\description{
Add necessary files for any R project
}
\examples{
withr::with_tempdir({
  project_path <- getwd()
  usethis::create_project(path = project_path, open = FALSE)
  add_kit_project(
    project_path,
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})
}
