% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to project}
\usage{
add_labels(project_id, lg = "fr")
}
\arguments{
\item{project_id}{project_id}

\item{lg}{by default, fr to get labels in french. Use en for english}
}
\value{
Table of added labels. Side effect: labels added on GitLab.
}
\description{
Add labels to project
}
\examples{
\dontrun{
add_labels(
  project_id = project_id
)
#'}
}
