% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wikis.R
\name{add_wikis}
\alias{add_wikis}
\title{Add Wiki}
\usage{
add_wikis(
  project_id,
  project_name,
  group_url,
  group = basename(group_url),
  type = c("home", "cr", "keys")
)
}
\arguments{
\item{project_id}{project_id}

\item{project_name}{project_name}

\item{group_url}{group_url}

\item{group}{group}

\item{type}{type}
}
\value{
Tibble with wikis and GitLab infos.
}
\description{
Add Wiki
}
\details{
Types:
\itemize{
\item home: Home Page
\item cr: Comptes-rendus
\item keys: Key dates of the project
}
}
\examples{
\dontrun{
add_wikis(
  project_id = project_id,
  project_name = project_name,
  group_url = group_url,
  group = basename(group_url)
)
}
}
