% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wikis_github.R
\name{add_wikis_github}
\alias{add_wikis_github}
\title{add_wikis_github}
\usage{
add_wikis_github(owner, repo, type = c("home", "cr", "keys", "weekly"))
}
\arguments{
\item{owner}{Owner of the repository}

\item{repo}{Repository name}

\item{type}{Wiki type to add}
}
\description{
add_wikis_github
}
\examples{
\dontrun{
add_wikis_github(
  owner = "ghowner",
  repo = "areponame"
)
}
}
