% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_yaml.R
\name{amend_yaml}
\alias{amend_yaml}
\title{Amend yaml}
\usage{
amend_yaml(path_to_yaml, connect_ci_list)
}
\arguments{
\item{path_to_yaml}{pat to .gtilab-ci.yml file}

\item{connect_ci_list}{list, output of \code{create_deploy_ci_stage}}
}
\description{
combien, if needed, to an existing .gitlab-ci.yml file
}
\examples{
# amend_yaml()
}
