% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4_book_template.R
\name{bs4_book_template}
\alias{bs4_book_template}
\title{Render bookdown bs4book}
\usage{
bs4_book_template(...)
}
\value{
A pagedown template
}
\description{
Render bookdown bs4book
}
\examples{
\dontrun{
bookdown::render_book("index.Rmd", output_format = "lozen::bs4_book_template")
}
}
