% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pkgdown_with_reports.R
\name{build_pkgdown_with_reports}
\alias{build_pkgdown_with_reports}
\title{Build a pkgdown with extra reports tab for testdown and coverage output}
\usage{
build_pkgdown_with_reports(
  pkg = ".",
  pkgdown_path = "public",
  assets_path = "pkgdown/assets",
  reports = c("coverage", "testdown", "gitdown"),
  git_branch_ref = "main",
  overwrite_assets = TRUE
)
}
\arguments{
\item{pkg}{character The path to the package to be build and reported}

\item{pkgdown_path}{character The relative path inside the package to store the final site}

\item{assets_path}{character The relative path within the package to store the reports outputs}

\item{reports}{character A vector of reports to be produced, can be any subset of \code{c("testdown","coverage")}}

\item{git_branch_ref}{character A vector of the git branch to use, 'main' by default}

\item{overwrite_assets}{logical Whether the assets directory to store reports should be overwritten}
}
\value{
None Generate a pkgdown with test and coverage reports
}
\description{
Build a pkgdown with extra reports tab for testdown and coverage output
}
\examples{
\dontrun{
# build_pkgdown_with_reports(
#    pkg = ".",
#    pkgdown_path = "public",
#    assets_path = "pkgdown/assets"),
#    reports = c("testdown","coverage")
}
}
