% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_yaml_exists.R
\name{check_if_yaml_exists}
\alias{check_if_yaml_exists}
\title{Check if yaml file exist}
\usage{
check_if_yaml_exists(dir = ".", file_name = ".gitlab-ci.yml", append = TRUE)
}
\arguments{
\item{dir}{directory to scan}

\item{file_name}{filename to use .gitlab-ci.yml}

\item{append}{boolean do we allow to append stage to existing ci file. TRUE by default}
}
\description{
Check if yaml file already exist and ask permission for overwriting
}
\examples{
# check_if_yaml_exists()
}
