% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_image.R
\name{clean_image}
\alias{clean_image}
\title{clean image}
\usage{
clean_image(ci)
}
\arguments{
\item{ci}{List of CI parameters, imported from yaml file}
}
\value{
list the list of CI parameters with a unique docker image
}
\description{
Manage image in yaml
}
\examples{
# pkgdown yml
full <- yaml::read_yaml(file = system.file("yaml", ".gitlab-ci-pkg.yml", package = "lozen"))

# shiny yaml
connect <- yaml::read_yaml(file = system.file("yaml", ".gitlab-ci-shiny.yml", package = "lozen"))

ci_list <- combine_ci(ci1 = full, ci2 = connect)
ci_list <- clean_image(ci_list)
}
