% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ci.R
\name{combine_ci}
\alias{combine_ci}
\title{combine CI}
\usage{
combine_ci(ci1, ci2)
}
\arguments{
\item{ci1}{List of initial CI parameters}

\item{ci2}{List of CI parameters to append to the initial CI}
}
\value{
list list of combined CI parameters
}
\description{
Merge two yaml as a list into a combined list
}
\examples{
# pkgdown yml
full <- yaml::read_yaml(file = system.file("yaml", ".gitlab-ci-pkg.yml", package = "lozen"))

# shiny yaml
connect <- yaml::read_yaml(file = system.file("yaml", ".gitlab-ci-shiny.yml", package = "lozen"))

ci_list <- combine_ci(ci1 = full, ci2 = connect)
}
