% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book_project.R
\name{create_book_project}
\alias{create_book_project}
\title{Transform project as book with lozen template}
\usage{
create_book_project(
  project_path,
  bookdown_path = system.file("lozendown", package = "lozen"),
  css = NULL,
  footer = NULL,
  logo = NULL,
  index = NULL,
  output_yml = NULL
)
}
\arguments{
\item{project_path}{project_path}

\item{bookdown_path}{a path to a bookdown}

\item{css}{path(s) for css to copy}

\item{footer}{path for footer to copy}

\item{logo}{path for logo to copy}

\item{index}{path for index to copy}

\item{output_yml}{path for output_yml to copy}
}
\value{
Side Effect: Transform project as book
}
\description{
Transform project as book with lozen template
}
\examples{
withr::with_tempdir({
  project_path <- getwd()
  create_book_project(project_path)
})
}
