% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deploy_ci_stage.R
\name{create_deploy_ci_stage}
\alias{create_deploy_ci_stage}
\title{Title}
\usage{
create_deploy_ci_stage(
  image,
  deploy_function,
  stage_name = deploy_function,
  ...
)
}
\arguments{
\item{image}{image}

\item{deploy_function}{deploy_function}

\item{stage_name}{stage_name}

\item{...}{dots not used}
}
\description{
Title
}
\examples{
create_deploy_ci_stage(
  image = "rocker/verse",
  deploy_function = "deploy_connect_shiny"
)
}
