% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_project.R
\name{create_group_project}
\alias{create_group_project}
\title{Create new project in a group}
\usage{
create_group_project(
  project_name,
  namespace_id,
  gitlab_con = "default",
  default_branch = "main"
)
}
\arguments{
\item{project_name}{name of the project to create}

\item{namespace_id}{id of the group in which to add the project}

\item{gitlab_con}{Connection credentials to GitLab.
Better use \code{set_gitlab_connection()} before this function}

\item{default_branch}{Default branch for the project created. Default to "main".}
}
\value{
project_id. Side effect: Create a project on GitLab if not exists.
}
\description{
Create new project in a group
}
\examples{
\dontrun{
create_group_project(
  project_name = project_name,
  namespace_id = namespace_id,
  default_branch = "main"
)
}
}
