% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{create_issue_content_clients}
\alias{create_issue_content_clients}
\title{Create issue content for client}
\usage{
create_issue_content_clients(project_name, group_url)
}
\arguments{
\item{project_name}{name of the project}

\item{group_url}{url to gitlab website}
}
\value{
character vector of the issue content
}
\description{
Generate text for client first issue in gitlab
}
\examples{
\dontrun{
create_issue_content_clients(
  project_name = "<get_your_project_name>",
  group_url = "<group_url_repo>" # should looks like "https://gitlab.com/cervan.girard/"
)
}
}
