% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{create_issue_content_clients_github}
\alias{create_issue_content_clients_github}
\title{Create issue content for client}
\usage{
create_issue_content_clients_github(owner, repo)
}
\arguments{
\item{owner}{owner of the project}

\item{repo}{name of the project}
}
\value{
character vector of the issue content
}
\description{
Generate text for client first issue in github
}
\examples{
\dontrun{
create_issue_content_clients_github(
  owner = "owner",
  repo = "repo"
)
}
}
