% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{create_issue_content_dev}
\alias{create_issue_content_dev}
\title{Create issue content for devs}
\usage{
create_issue_content_dev(project_id, group_url, project_name)
}
\arguments{
\item{project_id}{project_id}

\item{group_url}{group_url}

\item{project_name}{project_name}
}
\value{
character vector of the issue content
}
\description{
Generate text for devs first issue in gitlab
}
\examples{
\dontrun{
create_issue_content_dev(
  project_id = "<get_your_id_project>",
  project_name = "<get_your_project_name>",
  group_url = "<group_url_repo>" # should looks like "https://gitlab.com/cervan.girard/"
)
}
}
