% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_special_issues.R
\name{create_issue_content_kickoff}
\alias{create_issue_content_kickoff}
\title{Create issue content for kickoff}
\usage{
create_issue_content_kickoff()
}
\value{
character vector of the issue content
}
\description{
Generate text for kickoff issue
}
\examples{
create_issue_content_kickoff()
}
