% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_production.R
\name{create_production}
\alias{create_production}
\title{Create branch production from master}
\usage{
create_production(project_path)
}
\arguments{
\item{project_path}{project_path}
}
\value{
Side effect: New branch on the server
}
\description{
Create branch production from master
}
\examples{
\dontrun{
create_production(
  project_path = project_path
)
}
}
