% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_project.R
\name{create_r_project}
\alias{create_r_project}
\title{Create a R project with full content skeleton}
\usage{
create_r_project(
  project_path,
  type = c("package", "golem", "book"),
  name_licence,
  type_licence
)
}
\arguments{
\item{project_path}{project_path}

\item{type}{type}

\item{name_licence}{name for the licence}

\item{type_licence}{type of the licence, should be a function like usethis::use_mit_licence}
}
\value{
Side Effect: Transform project as package or book:
\itemize{
\item With package, golem or book skeleton
\item git necessary files
\item pkgdown or book template
\item GitLab CI
}
}
\description{
Create a R project with full content skeleton
}
\examples{
### R package skeleton
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("package"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})


### Shiny app
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("golem"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})

### Bookdown
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("book"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})

}
