% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_connect_bookdown.R
\name{deploy_connect_bookdown}
\alias{deploy_connect_bookdown}
\title{deploy_connect_bookdown}
\usage{
deploy_connect_bookdown(
  connect_url = Sys.getenv("CONNECT_URL"),
  connect_user = Sys.getenv("CONNECT_USER"),
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  app_name = Sys.getenv("APP_NAME", unset = Sys.getenv("CI_PROJECT_NAME")),
  deploy_dir = file.path(getwd(), "_book"),
  connect_name = "connect",
  file_to_ignore_regex =
    ".Rprofile$|^.Renviron$|renv/|rstudio_.*/|deliverables/|dev/|data-raw/|dockerfiles/",
  forceUpdate = FALSE,
  lint = FALSE,
  ...
)
}
\arguments{
\item{connect_url}{URL of the Connect server}

\item{connect_user}{User name to use to connect to the Connect server}

\item{connect_api_token}{API token to use to connect to the Connect server}

\item{app_name}{Name of the app to deploy}

\item{deploy_dir}{Directory to deploy}

\item{connect_name}{Name of the Connect server}

\item{file_to_ignore_regex}{Regex to use to ignore files}

\item{forceUpdate}{If \code{TRUE}, update any previously-deployed app without
asking. If \code{FALSE}, ask to update. If unset, defaults to the value of
\code{getOption("rsconnect.force.update.apps", FALSE)}.}

\item{lint}{Lint the project before initiating deployment, to identify
potentially problematic code?}

\item{...}{Other arguments to pass to rsconnect::deployApp}
}
\value{
used for side effects
}
\description{
Before using it, please follow these steps :
}
\details{
\itemize{
\item Ask the Mission Lead Dev for their deployment token on Connect, this is the one you will need to use.
\item Add the environment variables to your personal ".Renviron" to manually deploy to Connect:
\itemize{
\item Add \code{CONNECT_USER} with username.
\item Add \code{CONNECT_TOKEN} with the token.
\item Add \code{CONNECT_URL} with the connect url
}
}

TODO
}
\examples{
\dontrun{
if (
  Sys.getenv("CONNECT_URL") != "" &
    Sys.getenv("CONNECT_USER") != "" &
    Sys.getenv("CONNECT_TOKEN") != ""
) {
  project_name <- "lozen-example-bookdown"

  tmpdir <- tempfile(pattern = "book-")
  dir.create(tmpdir)
  project_path <- file.path(tmpdir, project_name)

  bookdown::create_bs4_book(path = project_path)

  bookdown::render_book(input = project_path)

  deploy_connect_bookdown(
    app_name = project_name,
    deploy_dir = file.path(project_path, "_book")
  )
}
}
}
