% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_connect_pkgdown.R
\name{deploy_connect_pkgdown}
\alias{deploy_connect_pkgdown}
\title{deploy_connect_pkgdown}
\usage{
deploy_connect_pkgdown(
  connect_url = Sys.getenv("CONNECT_URL"),
  connect_user = Sys.getenv("CONNECT_USER"),
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  app_name = NULL,
  deploy_dir = c(file.path(getwd(), "public"), file.path(getwd(), "docs"),
    file.path(getwd(), "inst/site/"), file.path(getwd(), ".")),
  connect_name = Sys.getenv("CONNECT_NAME", unset = "connect"),
  file_to_ignore_regex =
    ".Rprofile$|^.Renviron$|renv/|rstudio_.*/|deliverables/|dev/|data-raw/|dockerfiles/",
  forceUpdate = FALSE,
  lint = FALSE,
  app_primary_doc = "index.html",
  ...
)
}
\arguments{
\item{connect_url}{URL of the Connect server}

\item{connect_user}{User name to use to connect to the Connect server}

\item{connect_api_token}{API token to use to connect to the Connect server}

\item{app_name}{Name of the app to deploy}

\item{deploy_dir}{vectoro Directory path to deploy, the first available with be used.
default is c(file.path(getwd(), "public"),
file.path(getwd(), "docs"),
file.path(getwd(), "inst/site/"),
file.path(getwd(), ".")   )}

\item{connect_name}{Name of the Connect server}

\item{file_to_ignore_regex}{Regex to use to ignore files}

\item{forceUpdate}{If \code{TRUE}, update any previously-deployed app without
asking. If \code{FALSE}, ask to update. If unset, defaults to the value of
\code{getOption("rsconnect.force.update.apps", FALSE)}.}

\item{lint}{Lint the project before initiating deployment, to identify
potentially problematic code?}

\item{app_primary_doc}{If the application contains more than one document,
this parameter indicates the primary one, as a path relative to appDir.
Can be NULL, in which case the primary document is inferred from the
contents being deployed.}

\item{...}{Other arguments to pass to rsconnect::deployApp}
}
\value{
used for side effects
}
\description{
Before using it, please follow these steps :
}
\details{
\itemize{
\item Ask the Mission Lead Dev for their deployment token on Connect, this is the one you will need to use.
\item Add the environment variables to your personal ".Renviron" to manually deploy to Connect:
\itemize{
\item Add \code{CONNECT_USER} with username.
\item Add \code{CONNECT_TOKEN} with the token.
\item Add \code{CONNECT_URL} with the connect url
}
}

TODO
}
\examples{
\dontrun{
# We assume that you are working on a R package
# if not done yet, create your pkgdown
pkgdown::build_site(
  pkg = ".",
  override = list(destination = "inst/site/")
)
if (Sys.getenv("CONNECT_URL") != "" &
  Sys.getenv("CONNECT_USER") != "" &
  Sys.getenv("CONNECT_TOKEN") != "") {
  deploy_connect_pkgdown(
    app_name = "titi",
    deploy_dir = file.path(project_path, "inst/site/")
  )
}
}
}
