% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_connect_shiny.R
\name{deploy_connect_shiny}
\alias{deploy_connect_shiny}
\title{deploy_connect}
\usage{
deploy_connect_shiny(
  connect_url = Sys.getenv("CONNECT_URL"),
  connect_user = Sys.getenv("CONNECT_USER"),
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  app_name = NULL,
  deploy_dir = getwd(),
  connect_name = Sys.getenv("CONNECT_NAME", unset = "connect"),
  file_to_ignore_regex =
    ".Rprofile$|^.Renviron$|renv/|rstudio_.*/|deliverables/|dev/|data-raw/|dockerfiles/",
  forceUpdate = FALSE,
  lint = FALSE,
  ...
)
}
\arguments{
\item{connect_url}{URL of the Connect server}

\item{connect_user}{User name to use to connect to the Connect server}

\item{connect_api_token}{API token to use to connect to the Connect server}

\item{app_name}{Name of the app to deploy}

\item{deploy_dir}{Directory to deploy}

\item{connect_name}{Name of the Connect server}

\item{file_to_ignore_regex}{Regex to use to ignore files}

\item{forceUpdate}{What should happen if there's no deployment record for
the app, but there's an app with the same name on the server? If \code{TRUE},
will always update the previously-deployed app. If \code{FALSE}, will ask
the user what to do, or fail if not in an interactive context.

Defaults to \code{TRUE} when called automatically by the IDE, and \code{FALSE}
otherwise. You can override the default by setting option
\code{rsconnect.force.update.apps}.}

\item{lint}{Lint the project before initiating deployment, to identify
potentially problematic code?}

\item{...}{Other arguments to pass to rsconnect::deployApp}
}
\value{
used for side effects
}
\description{
Before using it, please follow these steps :
}
\details{
\itemize{
\item Ask the Mission Lead Dev for their deployment token on Connect, this is the one you will need to use.
\item Add the environment variables to your personal ".Renviron" to manually deploy to Connect:
\itemize{
\item Add \code{CONNECT_USER} with username.
\item Add \code{CONNECT_TOKEN} with the token.
\item Add \code{CONNECT_URL} with the connect url
}
}

TODO
}
\examples{
\dontrun{
deploy_connect_shiny(
  connect_url = Sys.getenv("CONNECT_URL"),
  connect_user = Sys.getenv("CONNECT_USER"),
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  app_name = "app_test"
)
}
}
