% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_connect.R
\name{fetch_connect}
\alias{fetch_connect}
\title{fetch_connect}
\usage{
fetch_connect(
  app_name,
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  connect_url = Sys.getenv("CONNECT_URL")
)
}
\arguments{
\item{app_name}{character Name of the app to be fetched}

\item{connect_api_token}{character API token to enable access to Connect}

\item{connect_url}{character URL to Connect server}
}
\value{
list A list containing the HTTP code status and the guid of the app
}
\description{
Fetch deployment status on Connect of a given app
}
\examples{
\dontrun{
fetch_connect(app_name = "my_app_name")
}
}
