% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_add_template_issue.R
\name{gh_add_template_issue}
\alias{gh_add_template_issue}
\title{Create the template for standard issue on gitlab}
\usage{
gh_add_template_issue(
  project_path = ".",
  language = c("fr", "en"),
  type = c("full", "light")
)
}
\arguments{
\item{project_path}{Path to the project}

\item{language}{Language. Should be one of 'fr' or 'en'}

\item{type}{Verbosity of comments in the template. Should be 'full' or 'light'.}
}
\value{
Side effect: create the template
}
\description{
Create the template for standard issue on gitlab
}
\examples{
# Try with a temp project
## Create temp dir and file
project_path <- tempfile("myproject")
dir.create(project_path)

## Add the full template in french
gh_add_template_issue(project_path, language = "fr", type = "full")

## Add the full template in english
gh_add_template_issue(project_path, language = "en", type = "full")

## Add the light template in french
gh_add_template_issue(project_path, language = "fr", type = "light")

## Add the light template in english
gh_add_template_issue(project_path, language = "en", type = "light")

# suppress tmp folder
unlink(project_path, recursive = TRUE)
}
