% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_create_weekly.R
\name{gh_create_weekly}
\alias{gh_create_weekly}
\title{Create a weekly issues summary for GitHub}
\usage{
gh_create_weekly(
  date_min = Sys.Date() - 7,
  date_max = Sys.Date(),
  user = "thinkr-open",
  repo = "fusen",
  verbose = FALSE,
  board_url
)
}
\arguments{
\item{date_min}{Minimal date to look for issues}

\item{date_max}{Maximal date to look for issues}

\item{user}{username or company name as shown on GitHub}

\item{repo}{GitHub repository}

\item{verbose}{Logical. Whether to return output in the console too.}

\item{board_url}{url of the GitHub board. If using the "classic" board, which is deprecated by GitHub, you may want to use this url: \code{glue("/repos/{user}/{repo}/projects")}}
}
\value{
A Weekly to copy-paste in a Wiki and a tibble
}
\description{
Create a weekly issues summary for GitHub
}
\examples{
\dontrun{
if (Sys.getenv("GITHUB_PAT") != "") {
  user <- "ThinkR-open"
  repo <- "example-weekly"
  weekly <- gh_create_weekly(
    date_min = "2022-06-30",
    date_max = "2022-06-30",
    user = user,
    repo = repo,
    board_url = glue::glue("/repos/{user}/{repo}/projects")
  )
}
}
# Copier dans le presse papier pour copier directement
# clipr::write_clip(weekly$weekly_info)
}
