% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_create_weekly_new_projects_board.R
\name{gh_create_weekly_new_projects_board}
\alias{gh_create_weekly_new_projects_board}
\title{Create a weekly issues summary for GitHub (new board)}
\usage{
gh_create_weekly_new_projects_board(
  date_min,
  date_max,
  user,
  repo,
  board_url,
  github_token = Sys.getenv("GITHUB_PAT"),
  verbose = FALSE
)
}
\arguments{
\item{date_min}{Minimal date to look for issues}

\item{date_max}{Maximal date to look for issues}

\item{user}{username or company name as shown on GitHub}

\item{repo}{GitHub repository}

\item{board_url}{url of the GitHub board.}

\item{github_token}{token access to graphQL API}

\item{verbose}{Logical. Whether to return output in the console too.}
}
\value{
A Weekly to copy-paste in a Wiki and a tibble
}
\description{
Create a weekly issues summary for GitHub (new board)
}
\examples{
\dontrun{
board_url <- "https://github.com/orgs/ThinkR-open/projects/4/"
date_min <- "2022-06-30"
date_max <- "2022-06-30"
user <- "ThinkR-open"
repo <- "example-weekly"

if (Sys.getenv("GITHUB_PAT") != "") {
  weekly <- gh_create_weekly_new_projects_board(
    date_min = date_min,
    date_max = date_max,
    user = user,
    repo = repo,
    board_url = board_url,
    github_token = Sys.getenv("GITHUB_PAT"),
    verbose = FALSE
  )
}
}
# Copier dans le presse papier pour copier directement
# clipr::write_clip(weekly$weekly_info)
}
