% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_add_template_issue.R
\name{gl_add_template_issue}
\alias{gl_add_template_issue}
\title{Create the template for standard issue on gitlab}
\usage{
gl_add_template_issue(
  project_path = ".",
  language = c("fr", "en"),
  type = c("full", "light")
)
}
\arguments{
\item{project_path}{Path to the project}

\item{language}{Language. Should be one of 'fr' or 'en'}

\item{type}{Verbosity of comments in the template. Should be 'full' or 'light'.}
}
\value{
Side effect: create the template
}
\description{
Create the template for standard issue on gitlab
}
\examples{
# Try with a temp project
## Create temp dir and file
project_path <- tempfile("myproject")
dir.create(project_path)

## Add the template in full french
gl_add_template_issue(project_path, language = "fr", type = "full")

## Add the template in full english
gl_add_template_issue(project_path, language = "en", type = "full")

## Add the template in light french
gl_add_template_issue(project_path, language = "fr", type = "light")

## Add the template in light english
gl_add_template_issue(project_path, language = "en", type = "light")

# Suppress tmp folder
unlink(project_path, recursive = TRUE)
}
