% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl_get_milestones_progress.R
\name{gl_get_milestones_progress}
\alias{gl_get_milestones_progress}
\title{Visualise the progress of the milestones for GitLab}
\usage{
gl_get_milestones_progress(
  project_id,
  language = c("fr", "en"),
  gitlab_url = Sys.getenv("GITLAB_URL", unset = "https://gitlab.com"),
  private_token = Sys.getenv("GITLAB_TOKEN"),
  color = "#f15522"
)
}
\arguments{
\item{project_id}{Id of the GitLab project}

\item{language}{Character. Language: 'fr' ou 'en'.}

\item{gitlab_url}{URL of the GitLab Forge}

\item{private_token}{Your private GitLab instance token allowed to use the "api".}

\item{color}{by defaut, warning color}
}
\value{
A ggplot
}
\description{
Visualise the progress of the milestones for GitLab
}
\examples{
\dontrun{
gl_get_milestones_progress(
  project_id = "<get_your_id_project>",
  private_token = Sys.getenv("GITLAB_TOKEN")
)
}
}
