% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql_to_tibble.R
\name{graphql_to_tibble}
\alias{graphql_to_tibble}
\title{Query GraphQL Github API}
\usage{
graphql_to_tibble(board_url, github_token = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{board_url}{url of the github project board}

\item{github_token}{access token to the graphql api}
}
\value{
a tibble
}
\description{
This function helps you retrieving the status of a project board within the organization
}
\examples{
\dontrun{

# Example with board hosted in an organization github account
board_url_organization <- "https://github.com/orgs/ThinkR-open/projects/4/"
github_token <- Sys.getenv("GITHUB_PAT")

graphql_to_tibble(
  board_url = board_url_organization,
  github_token = github_token
)


# Example with board hosted in a user github account
board_url_user <- "https://github.com/users/the-thinkr/projects/1"
github_token <- Sys.getenv("GITHUB_PAT")

graphql_to_tibble(
  board_url = board_url_user,
  github_token = github_token
)
}
}
