% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_odt.R
\name{html_to_odt}
\alias{html_to_odt}
\title{Convert html to odt with template}
\usage{
html_to_odt(input_html = "_main.html", output_odt = "_main.odt")
}
\arguments{
\item{input_html}{Path to html file to convert}

\item{output_odt}{Path to odt to create}
}
\description{
Convert html to odt with template
}
\examples{
\dontrun{
html_to_odt(input_html = "_main.html", output_odt = "_main.odt")
}
}
