% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lozen-package.R
\docType{package}
\name{lozen-package}
\alias{lozen}
\alias{lozen-package}
\title{lozen: Management tools for missions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Management tools for missions (internal and external). Includes weekly, GL projects, etc.
}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Authors:
\itemize{
  \item Margot Brard \email{margot@thinkr.fr} (\href{https://orcid.org/0000-0001-6754-0659}{ORCID})
  \item Murielle Delmotte \email{murielle@thinkr.fr} (\href{https://orcid.org/0000-0002-1339-2424}{ORCID})
  \item Swann Floc'hlay \email{swann@thinkr.fr} (\href{https://orcid.org/0000-0003-1477-830X}{ORCID})
  \item Yohann Mansiaux \email{yohann@thinkr.fr} (\href{https://orcid.org/0000-0002-8905-6603}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
