% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_autoclose_and_coverage.R
\name{modify_autoclose_and_coverage}
\alias{modify_autoclose_and_coverage}
\title{modify_autoclose_and_coverage on GitLab}
\usage{
modify_autoclose_and_coverage(
  project_id,
  autoclose = FALSE,
  build_coverage_regex = "Coverage: \\\\d+\\\\.\\\\d+"
)
}
\arguments{
\item{project_id}{project_id}

\item{autoclose}{Logical. Whether to autoclose issues when merged to main}

\item{build_coverage_regex}{Character. regex used to retrieve code coverage in Ci logs.}
}
\value{
Side Effect on GitLab
}
\description{
modify_autoclose_and_coverage on GitLab
}
\examples{
\dontrun{
modify_autoclose_and_coverage(
  project_id = project_id
)
}
}
