% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_issues_from_gitlab_to_github.R
\name{move_issues_from_gitlab_to_github}
\alias{move_issues_from_gitlab_to_github}
\title{Move Issues from GitLab to GitHub}
\usage{
move_issues_from_gitlab_to_github(
  gitlab_url = "https://gitlab.com/",
  github_url = "https://github.com/",
  gitlab_group,
  gitlab_repo,
  gitlab_project_id,
  github_owner,
  github_repo,
  gitlab_private_token = Sys.getenv("GITLAB_TOKEN"),
  github_token = NULL,
  sleep_time = 60,
  issue_start_over = 1,
  max_page = 10,
  force = FALSE
)
}
\arguments{
\item{gitlab_url}{GitLab instance URL}

\item{github_url}{GitHub URL, default to <github.com>}

\item{gitlab_group}{group name. Can be sub-group with "group/sub-group" format}

\item{gitlab_repo}{Repo name on GitLab}

\item{gitlab_project_id}{Repo ID on GitLab}

\item{github_owner}{Owner of the project on GitHub}

\item{github_repo}{Repo name of the project on GitHub. It should already exists.}

\item{gitlab_private_token}{GitLab Token with API permissions}

\item{github_token}{GitHub token. If NULL, will search for \code{GITHUB_PAT} in your env. variables}

\item{sleep_time}{Time in second to wait between each set of 5 issues. This is to reduce risk of being banned by GitHub. If you have time, Set to 60s by default.}

\item{issue_start_over}{For debug. Number of the issue in GitLab to start over with in case of temporary ban by GitHub.}

\item{max_page}{Maximum number of pages of issues to retrieve from GitLab.}

\item{force}{Whether to force writing issue to GitHub, even if issue numbers mismatch with GitLab.}
}
\value{
Used for side effects. Issues are dowloaded and copied to GitHub
}
\description{
Move Issues from GitLab to GitHub
}
\examples{
\dontrun{
move_issues_from_gitlab_to_github(
  gitlab_url = "https://gitlab.com/",
  github_url = "https://github.com/",
  gitlab_group = "thinkr-open",
  gitlab_repo = "example-weekly",
  gitlab_project_id = 37585948,
  github_owner = "ThinkR-open",
  github_repo = "example.mirror.issues.gl.gh",
  gitlab_private_token = Sys.getenv("GITLAB_TOKEN"),
  github_token = Sys.getenv("GITHUB_PAT"),
  sleep_time = 60,
  issue_start_over = 1,
  max_page = 10,
  force = FALSE
)
}
}
