% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_template.R
\name{paged_template}
\alias{paged_template}
\title{Render pagedown html_paged}
\usage{
paged_template(
  ...,
  css = c("default-fonts", "default-page", "default"),
  theme = NULL,
  template = pkg_resource("html", "paged.html"),
  csl = NULL,
  front_cover = NULL,
  back_cover = NULL
)
}
\arguments{
\item{...}{Arguments passed to
\code{bookdown::\link[bookdown]{html_document2}}.}

\item{css}{A character vector of CSS and Sass file paths. If a path
does not contain the \file{.css}, \file{.sass}, or \file{.scss} extension,
it is assumed to be a built-in CSS file. For example, \code{default-fonts}
means the file\code{pagedown:::pkg_resource('css', 'default-fonts.css')}.
To see all built-in CSS files, run \code{pagedown:::list_css()}.}

\item{theme}{The Bootstrap theme. By default, Bootstrap is not used.}

\item{template}{The path to the Pandoc template to convert Markdown to HTML.}

\item{csl}{The path of the Citation Style Language (CSL) file used to format
citations and references (see the \href{https://pandoc.org/MANUAL.html#citations}{Pandoc documentation}).}

\item{front_cover, back_cover}{Paths or urls to image files to be used
as front or back covers. Theses images are available through CSS variables
(see Details).}
}
\value{
A pagedown template
}
\description{
Render pagedown html_paged
}
