% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_branches.R
\name{protect_branches}
\alias{protect_branches}
\title{Protect 'main' and 'production' from push on server}
\usage{
protect_branches(project_id, unprotect = FALSE)
}
\arguments{
\item{project_id}{project_id}

\item{unprotect}{Logical. Whether to unprotect branches or not.}
}
\value{
Side effect: Branches protected from push on server.
}
\description{
Protect 'main' and 'production' from push on server
}
\examples{
\dontrun{
protect_branches(
  project_id = project_id
)
# unprotect if wanted
# protect_branches(project_id, unprotect = TRUE)
#'}
}
