% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_all_to_branch.R
\name{push_all_to_branch}
\alias{push_all_to_branch}
\title{Push everything on a branch}
\usage{
push_all_to_branch(
  project_path,
  branch = "main",
  main_branch = "main",
  message = "Init repo"
)
}
\arguments{
\item{project_path}{project_path}

\item{branch}{branch name. main by default}

\item{main_branch}{main branch name. main by default}

\item{message}{commit message}
}
\value{
Side effect: commit output
}
\description{
Push everything on a branch
}
\examples{
\dontrun{
push_all_to_branch(
  project_path = project_path
)
}
}
