% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_all_to_branch.R
\name{push_main}
\alias{push_main}
\title{Push main to server}
\usage{
push_main(
  project_path,
  branch = "main",
  main_branch = "main",
  message = "Init repo"
)
}
\arguments{
\item{project_path}{project_path}

\item{branch}{branch name. main by default}

\item{main_branch}{main branch name. main by default}

\item{message}{commit message}
}
\description{
Push main to server
}
