% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_all_to_branch.R
\name{push_master}
\alias{push_master}
\title{Push master to server}
\usage{
push_master(
  project_path,
  branch = "master",
  main_branch = "master",
  message = "Init repo"
)
}
\arguments{
\item{project_path}{project_path}

\item{branch}{branch name. main by default}

\item{main_branch}{main branch name. main by default}

\item{message}{commit message}
}
\description{
Push master to server
}
