% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ci.R
\name{read_ci}
\alias{read_ci}
\title{read_ci}
\usage{
read_ci(path = "./.gitlab-ci.yml")
}
\arguments{
\item{path}{Path to the yaml file}
}
\value{
list a list of all CI parameters
}
\description{
Import yaml CI files as list
}
\examples{
yaml_path <- system.file("yaml", ".gitlab-ci-shiny.yml", package = "lozen")
ci_list <- read_ci(path = yaml_path)
}
