% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci}
\alias{use_gitlab_ci}
\title{Set gitlab continuous integration}
\usage{
use_gitlab_ci(
  image = "rocker/verse",
  project_path = ".",
  type = "check-coverage-pkgdown",
  bookdown_output_format = c("lozen::paged_template", "lozen::bs4_book_template"),
  overwrite = TRUE
)
}
\arguments{
\item{image}{Docker image used as basis. See \url{https://github.com/rocker-org/rocker}}

\item{project_path}{Path of the project to add CI in.}

\item{type}{type of the CI template to use}

\item{bookdown_output_format}{If type="bookdown" it corresponds to the function used to output the bookdown}

\item{overwrite}{whether to overwrite existing GitLab CI yml file}
}
\description{
Set gitlab continuous integration
}
\details{
See \code{\link[gitlabr]{use_gitlab_ci}}
}
\examples{
withr::with_tempdir({
  use_gitlab_ci(image = "r-base")
})

withr::with_tempdir({
  use_gitlab_ci(
    image = "rocker/verse",
    type = "check-coverage-pkgdown"
  )
})
}
