% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_gitlab_project.R
\name{with_gitlab_project}
\alias{with_gitlab_project}
\title{Test gitlab CI workflows}
\usage{
with_gitlab_project(
  gitlab_url = Sys.getenv("GITLAB_URL", unset = "https://gitlab.com"),
  namespace_id,
  private_token = Sys.getenv("GITLAB_TOKEN"),
  connect_url = Sys.getenv("CONNECT_URL"),
  connect_api_token = Sys.getenv("CONNECT_TOKEN"),
  connect_user = Sys.getenv("CONNECT_USER"),
  connect_name = Sys.getenv("CONNECT_NAME", unset = "connect"),
  github_pat = Sys.getenv("GITHUB_PAT"),
  project_name = "lozen.test.project",
  branch_focus_for_ci = "main",
  exp
)
}
\arguments{
\item{gitlab_url}{A character string. The url of the gilab instance.}

\item{namespace_id}{An integer. The id of the gitlab group to
create the project in.}

\item{private_token}{A character string. The token to gain access to the
gitlab api.}

\item{connect_url}{A character string. The URL to the Connect server.}

\item{connect_api_token}{A character string. The token to gain access to the
Connect API. The token must be granted admin access.}

\item{connect_user}{A character string. User name on Connect.}

\item{connect_name}{A character string. Name of the Connect server (default "connect")}

\item{project_name}{A character string. The name of the project to be created.}

\item{branch_focus_for_ci}{Name of the branch to be targeted for the CI pipeline.}

\item{exp}{A valid R expression initializing a project and a gitlab ci
workflow.}

\item{github_pat.}{A character string. The token to gain access to the
github api.}
}
\value{
A data.frame containing information about jobs of
the gitlab project.
}
\description{
This functions allows its users to launch an
expression to initiate a given gitlab ci workflow within a dummy project on a #' gitlab instance.
}
\examples{
\dontrun{
output <- with_gitlab_project(
  gitlab_url = Sys.getenv("GITLAB_URL", unset = "https://gitlab.com"),
  namespace_id = NULL,
  project_name = "lozen.test.project",
  exp = {
    bookdown::create_bs4_book(path = ".")
    bookdown::render_book()
    lozen::use_gitlab_ci(
      type = "bookdown"
    )
  }
)
output
}
}
