% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_template.R
\name{paged_template}
\alias{paged_template}
\title{Render pagedown html_paged}
\usage{
paged_template(...)
}
\arguments{
\item{...}{Arguments passed to
\code{bookdown::\link[bookdown]{html_document2}}.}
}
\value{
A pagedown template
}
\description{
Render pagedown html_paged
}
\examples{
rmarkdown::render("my_file.Rmd", output_format = lozen::paged_template)
}
