% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_project.R, R/use_dev_history.R
\name{use_dev_history}
\alias{use_dev_history}
\title{Une a dev histroy file}
\usage{
use_dev_history(path = ".", type = c("package", "renv"))

use_dev_history(path = ".", type = c("package", "renv"))
}
\arguments{
\item{path}{Path to project to add dev_history in}

\item{type}{Type of dev_history. Multiple types possible among "package", "book", "renv".}
}
\description{
Une a dev histroy file

Une a dev histroy file
}
\examples{
withr::with_tempdir({
  use_dev_history(path = ".", type = c("package", "book", "renv"))
})
withr::with_tempdir({
  use_dev_history(path = ".", type = c("package", "book", "renv"))
})
}
