% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitlab_ci.R
\name{use_gitlab_ci}
\alias{use_gitlab_ci}
\title{Set gitlab continuous integration}
\usage{
use_gitlab_ci(
  image = "rocker/verse",
  repo_name = "https://packagemanager.rstudio.com/all/__linux__/focal/latest",
  project_path = ".",
  type = "check-coverage-pkgdown"
)
}
\arguments{
\item{image}{Docker image used as basis. See \url{https://github.com/rocker-org/rocker}}

\item{repo_name}{REPO_NAME environment variable for R CRAN mirror used}

\item{project_path}{Path of the project to add CI in.}

\item{type}{type of the CI template to use}
}
\description{
Set gitlab continuous integration
}
\details{
See \code{\link[gitlabr]{use_gitlab_ci}}
}
\examples{
withr::with_tempdir({
  use_gitlab_ci(image = "r-base")
})

withr::with_tempdir({
  use_gitlab_ci(
    image = "rocker/verse",
    repo_name = "https://packagemanager.rstudio.com/all/__linux__/focal/latest",
    type = "check-coverage-pkgdown"
  )
})
}
